\name{characteristicPolynomialRoots}
\alias{characteristicPolynomialRoots}
\title{Characteristic polynomial roots}
\usage{
  characteristicPolynomialRoots(theta_1, theta_2)
}
\arguments{
  \item{theta_1}{theta_1}

  \item{theta_2}{theta_2}
}
\description{
  No description
}
\details{
  No details.
}

