\name{eq3.011.alpha_t}
\alias{eq3.011.alpha_t}
\title{First-order affine dynamical system}
\usage{
  eq3.011.alpha_t(theta_0, theta_1, alpha_t.minus.1)
}
\arguments{
  \item{theta_0}{theta_0}

  \item{theta_1}{theta_1}

  \item{alpha_t.minus.1}{alpha_t.minus.1}
}
\description{
  Equation 3.11 - page 60
}
\details{
  No details.
}

