% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eq3.011.equilibrium}
\alias{eq3.011.equilibrium}
\title{First-order affine dynamical system equilibrium}
\usage{
eq3.011.equilibrium(theta_0, theta_1)
}
\arguments{
\item{theta_0}{theta_0}

\item{theta_1}{theta_1}
}
\description{
Equation in page 60 without a number
}
\details{
No details.
}

