% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eq3.012.alpha_t}
\alias{eq3.012.alpha_t}
\title{Simple linear dynamical system}
\usage{
eq3.012.alpha_t(alpha_t.minus.1)
}
\arguments{
\item{alpha_t.minus.1}{alpha_t.minus.1}
}
\description{
Equation 3.12 - page 61
}
\details{
No details.
}

