\name{eq3.019.alpha_t}
\alias{eq3.019.alpha_t}
\title{General solution of the second-order system when z1 != z2}
\usage{
  eq3.019.alpha_t(c_1, c_2, z_1, z_2, t)
}
\arguments{
  \item{c_1}{c_1}

  \item{c_2}{c_2}

  \item{z_1}{z_1}

  \item{z_2}{z_2}

  \item{t}{t}
}
\description{
  Equation 3.19 - page 63
}
\details{
  No details.
}

