% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eq3.020.c_2}
\alias{eq3.020.c_2}
\title{Constant c2 for the case 1 of the general solution of the second-order system}
\usage{
eq3.020.c_2(alpha_0, alpha_1, z_1, z_2)
}
\arguments{
\item{alpha_0}{alpha_0}

\item{alpha_1}{alpha_1}

\item{z_1}{z_1}

\item{z_2}{z_2}
}
\description{
Equation 3.20 - page 63
}
\details{
No details.
}

