\name{eq3.046.S_i}
\alias{eq3.046.S_i}
\title{Sensitivity matrix associated with the i-th latent root}
\usage{
  eq3.046.S_i(vhat_i, wapos_i, vapos_i, w_i)
}
\arguments{
  \item{vhat_i}{vhat_i}

  \item{wapos_i}{wapos_i}

  \item{vapos_i}{vapos_i}

  \item{w_i}{w_i}
}
\description{
  Equation 3.46 - page 69
}
\details{
  No details.
}

