% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eq3.049.alpha_t}
\alias{eq3.049.alpha_t}
\title{Logistic equation or logistic dynamical system}
\usage{
eq3.049.alpha_t(theta_0, theta_1, alpha_t.minus.1)
}
\arguments{
\item{theta_0}{theta_0}

\item{theta_1}{theta_1}

\item{alpha_t.minus.1}{alpha_t.minus.1}
}
\description{
Equation 3.49 - page 71
}
\details{
No details.
}

