\name{eq3.052.alpha_t}
\alias{eq3.052.alpha_t}
\title{Simple two-dimensional nonlinear system (alpha)}
\usage{
  eq3.052.alpha_t(alpha_t.minus.1, beta_t.minus.1, theta_1,
    theta_2)
}
\arguments{
  \item{theta_1}{theta_1}

  \item{theta_2}{theta_2}

  \item{alpha_t.minus.1}{alpha_t.minus.1}

  \item{beta_t.minus.1}{beta_t.minus.1}
}
\description{
  Equation 3.52 - page 76
}
\details{
  No details.
}

