% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eq3.053.beta_t}
\alias{eq3.053.beta_t}
\title{Simple two-dimensional nonlinear system (beta)}
\usage{
eq3.053.beta_t(alpha_t.minus.1, theta_2)
}
\arguments{
\item{theta_2}{theta_2}

\item{alpha_t.minus.1}{alpha_t.minus.1}
}
\description{
Equation 3.53 - page 76
}
\details{
No details.
}

