\name{eq3.081.f_of_omega}
\alias{eq3.081.f_of_omega}
\title{Spectral density for the Autoregressive (1) process}
\usage{
  eq3.081.f_of_omega(omega, var_w, alpha)
}
\arguments{
  \item{omega}{omega}

  \item{var_w}{var_w}

  \item{alpha}{alpha}
}
\description{
  Equation 3.81 - page 87
}
\details{
  No details.
}

