\name{eq3.082.Y_t}
\alias{eq3.082.Y_t}
\title{Autoregressive (2) process}
\usage{
  eq3.082.Y_t(alpha_1, alpha_2, Y_t.minus.1, Y_t.minus.2,
    W_t)
}
\arguments{
  \item{alpha_1}{alpha_1}

  \item{alpha_2}{alpha_2}

  \item{Y_t.minus.1}{Y_t.minus.1}

  \item{Y_t.minus.2}{Y_t.minus.2}

  \item{W_t}{W_t}
}
\description{
  Equation 3.82 - page 88
}
\details{
  No details.
}

