% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eq3.092.f_of_omega}
\alias{eq3.092.f_of_omega}
\title{Spectral density for Moving Average 1 MA(1) process}
\usage{
eq3.092.f_of_omega(omega, var_w, beta)
}
\arguments{
\item{omega}{omega}

\item{var_w}{var_w}

\item{beta}{beta}
}
\description{
Equation 3.92 - page 91
}
\details{
No details.
}

