\name{iterate.ar2}
\alias{iterate.ar2}
\title{Dummy for iterating equation 3.82 on page 90 to get figure 3.15}
\usage{
  iterate.ar2(n, alpha_1, alpha_2, W_t.vector)
}
\arguments{
  \item{n}{n}

  \item{alpha1}{alpha1}

  \item{alpha2}{alpha2}

  \item{W_t.vector}{W_t.vector}
}
\description{
  No description.
}
\details{
  No details.
}

