\name{iterate.chaos}
\alias{iterate.chaos}
\title{Dummy for iterating a function that takes its previous result as a parameter with a single initial parameter}
\usage{
  iterate.chaos(alpha_0, beta_0, theta_1, theta_2,
    iterations)
}
\arguments{
  \item{alpha_0}{alpha_0}

  \item{beta_0}{beta_0}

  \item{theta_1}{theta_1}

  \item{theta_2}{theta_2}

  \item{iterations}{iterations}
}
\description{
  No description.
}
\details{
  No details.
}

