\name{iterate.f1p}
\alias{iterate.f1p}
\title{Dummy for iterating a function that takes only one parameter (the previous result)}
\usage{
  iterate.f1p(f, alpha_0, iterations)
}
\arguments{
  \item{f}{function to iterate. It must be a function with
  one numeric input parameter and one numeric result (it
  uses the result of the last iteration as input for the
  next one)}

  \item{alpha_0}{Initial value}

  \item{iterations}{Number of iterations}
}
\description{
  No description.
}
\details{
  No details.
}

