\name{iterate.f3p}
\alias{iterate.f3p}
\title{Dummy for iterating a function that takes only two parameters and the previous result}
\usage{
  iterate.f3p(f, alpha_0, theta_0, theta_1, iterations)
}
\arguments{
  \item{f}{function to iterate. It must be a function with
  one numeri input parameter and one numeric result (it
  uses the result of the last iteration as input for the
  next one)}

  \item{alpha_0}{Initial value}

  \item{theta_0}{theta_0}

  \item{theta_1}{theta_1}

  \item{iterations}{Number of iterations}
}
\description{
  No description.
}
\details{
  No details.
}

