% Generated by roxygen2 (4.0.1): do not edit by hand
\name{randomWalk.gaussian}
\alias{randomWalk.gaussian}
\title{Gaussian random walk}
\usage{
randomWalk.gaussian(n, mean, sd)
}
\arguments{
\item{n}{number of desired samples}

\item{mean}{mean}

\item{sd}{standard deviation}
}
\description{
Example on page 86
}
\details{
No details.
}

