% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{rsstd}
\alias{rsstd}
\alias{rsstd-package}
\title{R scripts for the figures and examples from the book STATISTICS FOR SPATIO-TEMPORAL DATA by Cressie and Wikle 2011.}
\description{
rsstd provides R code for building the figures and examples in the book.
}
\references{
Noel Cressie, Christopher K. Wikle (2011). Statistics for Spatio-Temporal Data. Wiley. \url{http://eu.wiley.com/WileyCDA/WileyTitle/productCd-EHEP002348.html}.
}

