\name{sous.noAnalytic}
\alias{sous.noAnalytic}
\title{Non analytic solution of the second-order system}
\usage{
  sous.noAnalytic(f, alpha_0, alpha_1, iterations)
}
\arguments{
  \item{f}{A second-order system function like 3.30}

  \item{alpha_0}{alpha_0}

  \item{alpha_1}{alpha_1}

  \item{iterations}{iterations}
}
\description{
  Function for solving the second-order system on page 65.
  This function iterates the whole time series.
}
\details{
  No details.
}

