% Generated by roxygen2 (4.0.1): do not edit by hand
\name{whiteNoise.gaussian}
\alias{whiteNoise.gaussian}
\title{Gaussian white noise}
\usage{
whiteNoise.gaussian(n, mean, sd)
}
\arguments{
\item{n}{number of desired samples}

\item{mean}{mean}

\item{sd}{standard deviation}
}
\description{
Example on page 85
}
\details{
No details.
}

