% Generated by roxygen2 (4.0.1): do not edit by hand
\name{characteristicPolynomialRoots}
\alias{characteristicPolynomialRoots}
\title{Characteristic polynomial roots}
\usage{
characteristicPolynomialRoots(theta_1, theta_2)
}
\arguments{
\item{theta_1}{theta_1}

\item{theta_2}{theta_2}
}
\description{
No description
}
\details{
No details.
}

