% Generated by roxygen2 (4.0.2): do not edit by hand
\name{iterate.ar2}
\alias{iterate.ar2}
\title{Dummy for iterating equation 3.82 on page 90 to get figure 3.15}
\usage{
iterate.ar2(n, alpha_1, alpha_2, W_t.vector)
}
\arguments{
\item{n}{n}

\item{alpha_1}{alpha_1}

\item{alpha_2}{alpha_2}

\item{W_t.vector}{W_t.vector}
}
\description{
No description.
}
\details{
No details.
}

